extension_test/pyproject.toml
[tool.poetry]
name = "extension-test"
version = "0.1.0"
description = ""
authors = ["litmus-zhang <abdulsalamlukmon9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
amulet-leveldb = "^1.0.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
